tests: add unit test for .vbproj check
