+++ packages/react/src/__tests__/resources/.vs/VBTestProject/project-colors.json
@@ -1,0 +1,5 @@
+{
+  "Version": 1,
+  "ProjectMap": {},
+  "NextColorIndex": 0
+}
