+++ packages/react/src/__tests__/resources/VBTestProject.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32126.317
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "VBTestProject", "VBTestProject\VBTestProject.vbproj", "{199FF3DB-894E-42AD-A673-6162C73D332B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{199FF3DB-894E-42AD-A673-6162C73D332B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{199FF3DB-894E-42AD-A673-6162C73D332B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{199FF3DB-894E-42AD-A673-6162C73D332B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{199FF3DB-894E-42AD-A673-6162C73D332B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {221F2A12-EC56-42B2-9CEF-252E9F355506}
+	EndGlobalSection
+EndGlobal
